refactor auto instrument
