+++ lib/ddtrace/contrib/rails/auto_instrument_railtie.rb
@@ -1,0 +1,11 @@
+require 'ddtrace'
+
+# Railtie to include AutoInstrumentation in rails loading
+class Railtie < Rails::Railtie
+  # we want to load before config initializers so that any user supplied config
+  # in config/initializers/datadog.rb will take precedence
+  initializer 'datadog.start_tracer', before: :load_config_initializers do
+  	puts 'irun'
+    Datadog::Contrib::AutoInstrument.patch_all
+  end
+end
